refact(domain): move announcement into a separated namespace
